requirements.txt
fastapi==0.87.0
Pillow==9.3.0
uvicorn==0.19.0